reduce duplication for OpenIdProvider testing class
