+++ tests/test-distribution/test-ee10-distribution/pom.xml
@@ -62,6 +62,11 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.tests</groupId>
+      <artifactId>jetty-test-util</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.jetty.tests</groupId>
       <artifactId>jetty-testers</artifactId>
       <scope>test</scope>
     </dependency>
