+++ tests/test-distribution/test-ee11-distribution/src/test/java/org/eclipse/jetty/ee11/tests/distribution/OpenIdTests.java
@@ -17,8 +17,8 @@ import java.nio.file.Path;
 import java.util.concurrent.TimeUnit;
 
 import org.eclipse.jetty.client.ContentResponse;
-import org.eclipse.jetty.ee11.tests.distribution.openid.OpenIdProvider;
 import org.eclipse.jetty.http.HttpStatus;
+import org.eclipse.jetty.tests.OpenIdProvider;
 import org.eclipse.jetty.tests.distribution.AbstractJettyHomeTest;
 import org.eclipse.jetty.tests.testers.JettyHomeTester;
 import org.eclipse.jetty.tests.testers.Tester;
