+++ tests/test-distribution/test-ee9-distribution/pom.xml
@@ -58,6 +58,11 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.tests</groupId>
+      <artifactId>jetty-test-util</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.jetty.tests</groupId>
       <artifactId>jetty-testers</artifactId>
       <scope>test</scope>
     </dependency>
